Move states file
